% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gairEPA.R
\name{gairEPA}
\alias{gairEPA}
\title{Get US Air Data from the EPA website}
\usage{
gairEPA(year = 2012, type = "o3", freq = "daily", keepcsv = FALSE)
}
\arguments{
\item{year}{EPA offers data from 1990-2014, default is 2002}

\item{type}{o3, so2, co, no2, pm25, pm25a, pm10, pm25spec, wind, temp, bp, rhdp, haps, vocs, lead}

\item{freq}{daily, or hourly}

\item{keepcsv}{Whether to keep the csv data; default is FALSE}
}
\value{
A rda file in current directory
}
\description{
This function downloads air data for EPA.
}
\examples{
gairEPA(year = c(2013-2014), type = "so2")
gairEPA(2000, "co2", "hourly")
}
\keyword{air}
\keyword{pollution}

